requirements.txt
fastapi
pydantic
uvicorn
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
sqlalchemy
dotenv
pymongo
requests
python-dotenv
entsoe-py
